+++ esmerald/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.5.0"
+__version__ = "3.5.1"
 
 
 from lilya import status