+++ esmerald/config/static_files.py
@@ -26,6 +26,9 @@ class StaticFilesConfig(BaseModel):
     )
 
     app = Esmerald(static_files_config=static_files_config)
+    # or multiple in descending priority
+    app = Esmerald(static_files_config=[static_files_config1, static_files_config2, ...])
+
     ```
     """