+++ tests/databases/saffier/test_middleware_payload.py
@@ -6,7 +6,7 @@ from uuid import uuid4
 
 import pytest
 from anyio import from_thread, sleep, to_thread
-from httpx import AsyncClient
+from httpx import ASGITransport, AsyncClient
 from lilya.middleware import DefineMiddleware as LilyaMiddleware
 from pydantic import BaseModel
 from saffier.exceptions import DoesNotFound
@@ -190,7 +190,7 @@ def app():
 
 @pytest.fixture()
 async def async_client(app) -> AsyncGenerator:
-    async with AsyncClient(app=app, base_url="http://test") as ac:
+    async with AsyncClient(transport=ASGITransport(app=app), base_url="http://test") as ac:
         await to_thread.run_sync(blocking_function)
         yield ac